Fixed and reduced the output of `--sync-speed`.
When all the full node peers disconnect before the second measurement,
and the second measurement happens before the node status switches to
"Not Connected to Peers", the peak peer height in the second measurement
will be reported as `-1`. This is the internal error code for "Not
Connected to Peers".

The intended behaviour is to print "Not connected to peers." and exit.
To fix this, various checks are now performed during sync mode as well.

The change in peak peer height is also no longer reported on or used as a
reason to exit without print a speed or time remaining estimate, which
happens if the peak peer height is smaller in the second measurement.