Fixed `--sync-speed` Bug and Added More Warnings
When `--sync-speed` is ran while the full node is "Searching for an
initial chain" and "Not Connected to Peer(s)", instead of printing
"Searching for an initial chain" and exiting, it will print
"Not synced. Height: -2/-1". This commit fixes this.

When there are no blocks synced yet and no peers connected, the messages
"There is no blockchain found yet." and "Searching for an initial chain"
will be accompanied with "Additionally, no peers are connected.".