Update `chiavdf` to 1.1.1
Installing the timelord on an Ubuntu system (such as Ubuntu 22) seems to
result in the error `cannot import name 'setuptools' from 'setuptools'`.
Changing the `chiavdf` version in `setup.py` to 1.1.1 allows
`install-timelord.py` to complete, albeit with an error saying it's not
officially compatible. The same workaround works for Chia 1.5.0 and
should no longer be necessary in the next STAI update.